avengers_django_models/requirements.txt
django==2.0.2
django-debug-toolbar
mongoengine==0.9
djongo
pymongo>=3.2.0
django-rest-framework-mongoengine
setuptools
Pillow
jsonfield>=2.0.2
djangorestframework
